Clean up construct request from wsgi environ
