+++ ripozo/resources/request.py
@@ -6,178 +6,11 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from cgi import parse_header
-import json
 import logging
-import warnings
-
-import six
-from six.moves import urllib
 
 from ripozo.resources.constants import input_categories
 
 _LOG = logging.getLogger(__name__)
-
-
-class _Headers(dict):
-    """"
-    A case insensitive dictionary for headers.
-    Includes a helper factory method ``from_wsgi_environ``
-    for easy construction from a wsgi environ object
-    """
-    def __setitem__(self, key, value):
-        """
-        Case insensitive set item method
-
-        :param unicode key:
-        :param unicode value:
-        """
-        super(_Headers, self).__setitem__(key.title(), value)
-
-    def __getitem__(self, key):
-        """
-        Case insensitive __getitem__ implementation
-
-        :param unicode key:
-        :rtype: unicode
-        """
-        return super(_Headers, self).__getitem__(key.title())
-
-    @classmethod
-    def from_wsgi_environ(cls, environ):
-        """
-        Constructs a Headers dictionary from
-        a wsgi environ object
-
-        :param dict environ: The wsgi environ object
-        :type environ:
-        :return: A case insentive dictionary of valid headers
-        :rtype: _Headers
-        """
-        headers = cls()
-        for key, value in six.iteritems(environ):
-            if key.startswith('HTTP_'):
-                key = key[5:]
-                key = key.replace('_', '-')
-                headers[key] = value
-            if key in ['CONTENT_TYPE', 'CONTENT_LENGTH']:
-                key = key.replace('_', '-')
-                headers[key] = value
-        return headers
-
-
-def parse_form_encoded(form_encoded_string, charset='utf-8'):
-    """
-    Safely parses a form encoded string.  Starting
-    in ripozo v2.0.0 bad query strings will raise an
-    Exception
-
-    :param unicode form_encoded_string:
-    :return: The parsed values from the string
-    :rtype: dict{unicode: list[unicode]}
-    """
-
-    try:
-        query_parts = urllib.parse.parse_qs(form_encoded_string,
-                                            keep_blank_values=True,
-                                            strict_parsing=True)
-    except ValueError as e:
-        warnings.warn('The query string "{0}" is invalid.  In '
-                      'ripozo v2.0.0 this will raise a ValidationException. '
-                      'Error Message: {1}'.format(form_encoded_string, e),
-                      DeprecationWarning)
-        query_parts = urllib.parse.parse_qs(form_encoded_string,
-                                            keep_blank_values=True,
-                                            strict_parsing=False)
-
-    # Already unicode return as is
-    if six.PY3:
-        return query_parts
-
-    # Convert to unicode
-    unicode_query_parts = {}
-    for key, value in six.iteritems(query_parts):
-        unicode_query_parts[key.decode(charset)] = [val.decode(charset) for val in value]
-    return unicode_query_parts
-
-
-def _parse_query_string(environ):
-    """
-    Retrieves the query string from the wsgi
-    environ object and returns the parsed values
-
-    :param dict environ: The wsgi environ
-    :return: The parsed values from the query string
-    :rtype: dict
-    """
-    query_string = environ.get('QUERY_STRING', '')
-    return parse_form_encoded(query_string)
-
-
-def _get_charset(environ):
-    """
-    Gets the character set from the wsgi environ
-    via the Content-Type header.
-
-    :param dict environ: The WSGI environ object
-    :return: The string identifying the character
-        set of the request
-    :rtype: unicode
-    """
-    content_type = environ.get('CONTENT_TYPE', '')
-    # Decode according to RFC 5987 https://tools.ietf.org/html/rfc5987
-    if not isinstance(content_type, six.text_type):
-        content_type = content_type.decode('latin1')
-    content_type, params = parse_header(content_type)
-    return params.get('charset', 'utf-8')
-
-
-def parse_body(environ):
-    """
-    Parse the request body by first attempting to load it
-    as JSON and then trying to parse it as a form encoded
-    string
-
-    :param dict environ: The wsgi environ dictionary
-    :return: The parsed body
-    :rtype: dict
-    """
-    unicode_body = coerce_body_to_unicode(environ)
-    return parse_raw_body(unicode_body)
-
-
-def coerce_body_to_unicode(environ):
-    raw_body_file = environ.get('wsgi.input')
-    raw_body = raw_body_file.read() if raw_body_file else ''
-
-    # Decode the body into unicode
-    if not isinstance(raw_body, six.text_type):
-        charset = _get_charset(environ)
-        raw_body = raw_body.decode(charset)  # TODO this should appropriately raise an error
-    return raw_body
-
-
-def parse_raw_body(body):
-    # TODO docstring
-    if not body:
-        return {}
-    try:
-        return json.loads(body)
-    except ValueError:
-        return parse_form_encoded(body)
-
-
-def json_loads_backwards_compatible(body, content_type):
-    try:
-        return json.loads(body)
-    except ValueError:
-        warnings.warn('The content type "{}" only allows for '
-                      'json requests bodies.  The automatic handling '
-                      'of form encoded bodies will be deprecated in '
-                      'ripozo v2.0.0. In the future, an exception '
-                      'will be raised and the client will receive a '
-                      '4xx response '.format(content_type), DeprecationWarning)
-        return parse_form_encoded(body)
 
 
 class RequestContainer(object):
@@ -210,20 +43,6 @@ class RequestContainer(object):
         self._headers = headers or {}
         self.environ = environ or {}
         self.method = method
-
-    @classmethod
-    def from_wsgi_environ(cls, environ, url_params):
-        headers = _Headers.from_wsgi_environ(environ)
-        query_args = _parse_query_string(environ)
-        body_args = parse_body(environ)
-        return cls(
-            headers=headers,
-            query_args=query_args,
-            body_args=body_args,
-            url_params=url_params,
-            method=environ['REQUEST_METHOD'],
-            environ=environ
-        )
 
     @property
     def url_params(self):
