+++ ripozo/wsgi/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
